refactor(#151): split oversized modules into focused submodules
mcp-server/src/lib.rs (3695 → 3374 lines):
- Extract metering types/fns into metering.rs (PlanTier, UsageReport,
  UsageMeters, estimate_usage_cost, SkillUsageDelta, etc.)
- Extract JSON-RPC protocol types/fns into protocol.rs (JsonRpcRequest,
  JsonRpcResponse, negotiate_protocol_version, sanitize_log_field, etc.)
- Extract external DTO structs into dto.rs (ExternalSemanticState,
  StateMetadata, ExternalInteractiveElement)
- Re-export public API via pub use to preserve existing crate interface

core-runtime/src/browser.rs (2995 lines) → browser/mod.rs + helpers.rs:
- Convert browser.rs to browser/mod.rs module directory
- Extract 20 pure utility functions into browser/helpers.rs (pub(super))
  including tree traversal, key lookup, timeout math, error classification,
  path indexing, bbox conversion, and epoch timestamp helpers
- STABLE_KEY_SHORT_LEN const remains pub in mod.rs for re-export

No behavior changes; public API surfaces unchanged.
