+++ mcp-server/src/dto.rs
@@ -1,0 +1,37 @@
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
+use std::collections::BTreeMap;
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct ExternalSemanticState {
+    pub metadata: StateMetadata,
+    pub interactive_elements: Vec<ExternalInteractiveElement>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct StateMetadata {
+    pub url: String,
+    pub page_instance_id: String,
+    pub state_hash: String,
+    pub load_profile: String,
+    pub timestamp: u64,
+    /// `true` when this state was served from a verified speculative
+    /// pre-generation rather than a fresh capture (Spec §3.5 / ISSUE-147).
+    #[serde(default)]
+    pub speculative: bool,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct ExternalInteractiveElement {
+    pub id: i64,
+    pub stable_key: String,
+    pub alias: String,
+    pub role: String,
+    pub name: String,
+    pub attributes: BTreeMap<String, Value>,
+    pub bbox: [f64; 4],
+    pub policy_flags: Vec<String>,
+    /// Prompt-injection security classification flags (omitted when empty for backward compat).
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub security_flags: Vec<String>,
+}
